ext_man: Move extended manifest generator out of kernel folder
Generator is part of rimage tool and there is any reference to them
in kernel source code, so it should be moved out of kernel folder.
