+++ src/include/rimage/ext_manifest_gen.h
